requirements.txt
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
locustio==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack-python==0.5.6
netifaces==0.10.6
pylint==2.4.3
python-engineio==3.10.0
python-socketio==4.3.1
pyzmq==18.1.0
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
